requirements.txt
# Core dependencies
fastapi==0.110.2
uvicorn==0.29.0
httpx==0.27.0
tenacity==8.2.3
python-dotenv==1.0.1
pydantic==2.7.1
langgraph==0.0.18
chromadb==0.4.24
anthropic==0.21.3
openai==1.30.1

# Logging
loguru==0.7.2

# Tokenization (for chunking)
tiktoken==0.6.0

# Typing
typing-extensions==4.11.0

# For async file I/O (if needed)
aiofiles==23.2.1
